some fixes
fixed some bugs and erros but other than that the handling multiple dollar signs
$$$HOME
the previous algorithm wasnt working properly so I need to change it. rather than keeping the search from the last dollar sign going to use the location of the last dollar sign as a limiter for to search and double check if the current sign is at the previous index of the last dollar sign if so I will skip that dollar sign and update the last dollar sign var. yes yes I will do it next time coming soon